+++ src/renderer.rs
@@ -1,4 +1,3 @@
-use tar;
 use futures::sync::oneshot;
 use futures::future;
 use futures::Future;
@@ -13,35 +12,17 @@ use std::process::Command;
 use tokio_core::reactor::Handle;
 use tokio_process::CommandExt;
 use tera::Tera;
-use chrono::Utc;
-use rusoto;
-use serde_json;
-use s3;
-use s3::S3;
 use slog::{Drain, Duplicate, Logger};
 use sloggers::Build;
 use sloggers::types::Severity;
 use sloggers::file::FileLoggerBuilder;
-use std::default::Default;
-use std::fs::File;
-use error_chain::ChainedError;
-use mime;
 
 use http::*;
-use papers::{DocumentSpec, PapersUri, Summary};
+use papers::{DocumentSpec, PapersUri};
 use error::{Error, ErrorKind};
 use config::Config;
-
-fn extract_filename_from_uri(uri: &Uri) -> Option<String> {
-    match uri.path().split('/').last() {
-        Some(name) if !name.is_empty() => Some(name.to_string()),
-        _ => None,
-    }
-}
-
-fn s3_dir_name() -> String {
-    format!("{}", Utc::now())
-}
+use utils::s3::*;
+use utils::callbacks::*;
 
 #[derive(Debug)]
 pub struct Renderer<S>
@@ -352,196 +333,4 @@ where
         })
     });
     Box::new(future::join_all(futures))
-}
-
-/// This reports to the provided callback url with the presigned URL of the generated PDF and the
-/// location of the debugging output. It returns the response from the callback url as a future.
-fn report_success<S>(
-    config: &'static Config,
-    logger: Logger,
-    client: S,
-    callback_url: Uri,
-    s3_prefix: String,
-    presigned_url: String,
-) -> Box<Future<Item = (), Error = Error>>
-where
-    S: Service<Request = Request, Response = Response, Error = hyper::Error> + 'static + Clone,
-{
-    let outcome = Summary::File {
-        file: presigned_url,
-        s3_folder: s3_prefix,
-    };
-
-    debug!(logger, "Summary sent to callback: {:?}", outcome);
-
-    let callback_response = future::result(serde_json::to_vec(&outcome))
-        .map_err(|err| {
-            Error::with_chain(err, "Error encoding the rendering outcome")
-        })
-        .and_then(move |body| {
-            let req = Request::new(hyper::Method::Post, callback_url).with_body(body.into())
-                .with_header(hyper::header::ContentType(mime::APPLICATION_JSON));
-
-            client.call(req).map_err(|err| {
-                Error::with_chain(err, "Error posting to callback URL")
-            })
-        });
-
-    let response_bytes = {
-        let logger = logger.clone();
-        let max_asset_size = config.max_asset_size.clone();
-        callback_response.and_then(move |response| {
-            info!(
-                logger,
-                "Callback response: {}",
-                response.status().canonical_reason().unwrap_or("unknown")
-            );
-
-            response.get_body_bytes_with_limit(max_asset_size)
-        })
-    };
-
-    Box::new({
-        let logger = logger.clone();
-        response_bytes.and_then(move |bytes| {
-            debug!(
-                logger,
-                "Callback response body: {:?}",
-                ::std::str::from_utf8(&bytes).unwrap_or("<binary content>")
-            );
-            future::ok(())
-        })
-    })
-}
-
-/// When an error occurs during the generation process, it is reported with this function. It calls
-/// the callback_url from the document spec, posting a `Summary` object with the error and the key
-/// where the debug output can be found.
-fn report_failure<S>(
-    logger: Logger,
-    client: S,
-    error: Error,
-    s3_prefix: String,
-    callback_url: Uri,
-) -> Box<Future<Item = (), Error = Error>>
-where
-    S: Service<Request = Request, Response = Response, Error = hyper::Error> + 'static,
-{
-    error!(logger, "Reporting error: {}", error.display_chain());
-    let outcome = Summary::Error {
-        error: format!("{}", error),
-        s3_folder: s3_prefix,
-    };
-    debug!(logger, "Summary sent to callback: {:?}", outcome);
-    let res = future::result(serde_json::to_vec(&outcome))
-        .map_err(Error::from)
-        .and_then(move |body| {
-            let req = Request::new(hyper::Method::Post, callback_url).with_body(body.into())
-                .with_header(hyper::header::ContentType(mime::APPLICATION_JSON));
-            client.call(req).map_err(Error::from)
-        })
-        .map(|_| ());
-    Box::new(res)
-}
-
-/// Posts the file to the given key in S3 with the default S3 configuration.
-fn post_to_s3(config: &'static Config, path: &Path, key: String) -> Result<(), Error> {
-    let client = s3::S3Client::new(
-        rusoto::request::default_tls_client().expect("could not create TLS client"),
-        config,
-        config.s3.region,
-    );
-    debug!(config.logger, "Uploading {:?} to {:?}", path, key);
-    let mut body: Vec<u8> = Vec::new();
-    let mut file = File::open(path)?;
-    file.read_to_end(&mut body)?;
-    let request = s3::PutObjectRequest {
-        body: Some(body),
-        bucket: config.s3.bucket.clone(),
-        key,
-        ..Default::default()
-    };
-
-    client
-        .put_object(&request)
-        .map(|_| ())
-        .map_err(|err| Error::with_chain(err, "Error during S3 upload"))
-}
-
-/// Gets a presigned url for the specified key in the bucket specified in the configuration.
-///
-/// This does not perform any request.
-fn get_presigned_url(config: &'static Config, key: String) -> Result<String, Error> {
-    let client = s3::S3Client::new(
-        rusoto::request::default_tls_client().expect("could not create TLS client"),
-        config,
-        config.s3.region,
-    );
-    let request = s3::GetObjectRequest {
-        bucket: config.s3.bucket.clone(),
-        key,
-        response_expires: Some(format!("{}", config.s3.expiration_time)),
-        ..Default::default()
-    };
-    client.presigned_url(&request).map_err(|err| {
-        Error::with_chain(err, "Could not generate presigned url")
-    })
-}
-
-/// This function is responsible for uploading a tar file with the contents from the workspace (the
-/// temporary directory where we generated the PDF) to S3 under the given key.
-fn upload_workspace(
-    config: &'static Config,
-    logger: Logger,
-    workspace: PathBuf,
-    key: String,
-) -> Result<(), Error> {
-    debug!(logger, "Tarring {:?}", workspace);
-    let mut tarred_workspace: Vec<u8> = Vec::new();
-    {
-        let dir_name: PathBuf = workspace
-            .clone()
-            .components()
-            .last()
-            .unwrap()
-            .as_os_str()
-            .into();
-        debug!(logger, "tar {:?} as {:?}", &workspace, &dir_name);
-        let mut tarrer = tar::Builder::new(&mut tarred_workspace);
-        tarrer.append_dir_all(&dir_name, &workspace)?;
-        debug!(logger, "Tar was successful");
-        tarrer.finish()?;
-    }
-
-    // Write the tarred workspace to disk
-    let mut tar_file_path = workspace.to_path_buf();
-    tar_file_path.push("workspace.tar");
-    let mut output_file = File::create(&tar_file_path)?;
-    output_file.write_all(&tarred_workspace)?;
-
-    // Upload the tarred workspace to S3
-    post_to_s3(config, &tar_file_path, key)
-}
-
-#[cfg(test)]
-mod tests {
-    use hyper::Uri;
-    use super::extract_filename_from_uri;
-
-    #[test]
-    fn test_extract_filename_from_uri_works() {
-        let assert_extracted = |input: &'static str, expected_output: Option<&'static str>| {
-            let uri = input.parse::<Uri>().unwrap();
-            assert_eq!(
-                extract_filename_from_uri(&uri),
-                expected_output.map(|o| o.to_string())
-            );
-        };
-
-        assert_extracted("/logo.png", Some("logo.png"));
-        assert_extracted("/assets/", None);
-        assert_extracted("/assets/icon", Some("icon"));
-        assert_extracted("/", None);
-        assert_extracted("http://www.store2be.com", None);
-    }
 }